Refactor Task Service.
