+++ src/Service/Task/TaskService.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Service;
+namespace App\Service\Task;
 
 use App\Exception\Task;
 use App\Repository\TaskRepository;
+use App\Service\BaseService;
+use App\Service\RedisService;
 
 final class TaskService extends BaseService
 {
